requirements.txt
lida
lida[transformers]
torch
streamlit

langchain==0.0.146
python-dotenv==1.0.0
streamlit== 1.22.0
openai
Tabulate 